fix(storybook): various fixes to the storybook configuration
