+++ packages/patterns-react/.storybook/addons.js
@@ -1,9 +1,7 @@
-import '@storybook/addon-storysource/register';
 import '@storybook/addon-knobs/register';
-import '@storybook/addon-actions/register';
-import '@storybook/addon-links/register';
-import '@storybook/addon-a11y/register';
-import '@storybook/addon-options/register';
-
-// Community addons
 import 'storybook-readme/register';
+// import '@storybook/addon-actions/register';
+import '@storybook/addon-storysource/register';
+import '@storybook/addon-links/register';
+// import '@storybook/addon-a11y/register';
+import '@storybook/addon-options/register';
